Rename init package to nvm
