+++ subprojects/distributions-management/src/functionalTest/java/dev/nokee/nvm/NokeeVersionManagementServiceNoVersionFunctionalTest.java
@@ -13,13 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.init;
+package dev.nokee.nvm;
 
 import dev.gradleplugins.runnerkit.BuildResult;
 import dev.gradleplugins.runnerkit.GradleRunner;
 import dev.nokee.internal.testing.junit.jupiter.ContextualGradleRunnerParameterResolver;
-import dev.nokee.internal.testing.junit.jupiter.GradleFeatureRequirement;
-import dev.nokee.internal.testing.junit.jupiter.RequiresGradleFeature;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectory;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectoryExtension;
 import org.junit.jupiter.api.BeforeEach;
@@ -32,40 +30,32 @@ import java.nio.file.Path;
 import java.util.Arrays;
 
 import static dev.gradleplugins.buildscript.blocks.PluginsBlock.plugins;
-import static dev.nokee.init.fixtures.DotNokeeVersionTestUtils.writeVersionFileTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.not;
 
-@RequiresGradleFeature(GradleFeatureRequirement.CONFIGURATION_CACHE)
 @ExtendWith({TestDirectoryExtension.class, ContextualGradleRunnerParameterResolver.class})
-class ConfigurationCacheDetectsChangesToDotNokeeRCFileFunctionalTest {
+class NokeeVersionManagementServiceNoVersionFunctionalTest {
 	@TestDirectory Path testDirectory;
 	GradleRunner executer;
-	BuildResult result;
 
 	@BeforeEach
 	void setup(GradleRunner runner) throws IOException {
-		writeVersionFileTo(testDirectory, "0.3.0");
+		executer = runner;
 		plugins(it -> it.id("dev.nokee.nokee-version-management")).writeTo(testDirectory.resolve("settings.gradle"));
-		executer = runner.withArgument("verify").withArgument("--configuration-cache");
 		Files.write(testDirectory.resolve("build.gradle"), Arrays.asList(
-			"plugins {",
-			"  id 'dev.nokee.jni-library'", // we must resolve a Nokee plugin so the "version" is marked as used
-			"}",
-			"tasks.register('verify')"
+			"def service = gradle.sharedServices.registrations.nokeeVersionManagement.service",
+			"tasks.register('verify') {",
+			"  usesService(service)",
+			"  doLast {",
+			"    service.get().version",
+			"  }",
+			"}"
 		));
-		result = executer.build();
 	}
 
 	@Test
-	void reusesConfigurationCacheWhenVersionFileDoesNotChanges() {
-		assertThat(executer.build().getOutput(), containsString("Reusing configuration cache"));
-	}
-
-	@Test
-	void doesNotReuseConfigurationCacheWhenVersionFileChange() throws IOException {
-		writeVersionFileTo(testDirectory, "0.4.0");
-		assertThat(executer.build().getOutput(), not(containsString("Reusing configuration cache")));
+	void showsMeaningfulError() {
+		final BuildResult result = executer.withTasks("verify").buildAndFail();
+		assertThat(result.getOutput(), containsString("Please add the Nokee version to use in a .nokee-version file."));
 	}
 }
