+++ subprojects/distributions-management/src/functionalTest/java/dev/nokee/nvm/NokeeVersionManagementServiceUsesVersionFromBuildSrcParentFunctionalTest.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.init;
+package dev.nokee.nvm;
 
-import dev.gradleplugins.buildscript.blocks.SettingsBlock;
 import dev.gradleplugins.runnerkit.GradleRunner;
 import dev.nokee.internal.testing.junit.jupiter.ContextualGradleRunnerParameterResolver;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectory;
@@ -30,42 +29,38 @@ import java.nio.file.Path;
 import java.util.Arrays;
 
 import static dev.gradleplugins.buildscript.blocks.PluginsBlock.plugins;
-import static dev.nokee.init.fixtures.DotNokeeVersionTestUtils.writeVersionFileTo;
+import static dev.nokee.nvm.fixtures.DotNokeeVersionTestUtils.writeVersionFileTo;
 
 @ExtendWith({TestDirectoryExtension.class, ContextualGradleRunnerParameterResolver.class})
-class NokeeVersionManagementServiceUsesVersionFromIncludedBuildParentFunctionalTest {
+class NokeeVersionManagementServiceUsesVersionFromBuildSrcParentFunctionalTest {
 	@TestDirectory Path testDirectory;
 	GradleRunner executer;
 
 	@BeforeEach
 	void setup(GradleRunner runner) throws IOException {
 		executer = runner;
-		SettingsBlock.builder().plugins(it -> it.id("dev.nokee.nokee-version-management"))
-			.includeBuild("build-src")
-			.build()
-			.writeTo(testDirectory.resolve("settings.gradle"));
+		plugins(it -> it.id("dev.nokee.nokee-version-management")).writeTo(testDirectory.resolve("settings.gradle"));
 		writeVersionFileTo(testDirectory, "0.4.2");
 		Files.write(testDirectory.resolve("build.gradle"), Arrays.asList(
-			"tasks.register('verify') {",
-			"  dependsOn gradle.includedBuild('build-src').task(':verify')",
-			"}"
+			"tasks.register('verify')"
 		));
 
-		Files.createDirectory(testDirectory.resolve("build-src"));
-		plugins(it -> it.id("dev.nokee.nokee-version-management")).writeTo(testDirectory.resolve("build-src/settings.gradle"));
-		Files.write(testDirectory.resolve("build-src/build.gradle"), Arrays.asList(
+		Files.createDirectory(testDirectory.resolve("buildSrc"));
+		plugins(it -> it.id("dev.nokee.nokee-version-management")).writeTo(testDirectory.resolve("buildSrc/settings.gradle"));
+		Files.write(testDirectory.resolve("buildSrc/build.gradle"), Arrays.asList(
 			"def service = gradle.sharedServices.registrations.nokeeVersionManagement.service",
-			"tasks.register('verify') {",
+			"def verifyTask = tasks.register('verify') {",
 			"  usesService(service)",
 			"  doLast {",
 			"    assert service.get().version.toString() == '0.4.2'",
 			"  }",
-			"}"
+			"}",
+			"tasks.named('build') { finalizedBy(verifyTask) }"
 		));
 	}
 
 	@Test
-	void fetchesNokeeVersionFromParentBuild() {
+	void fetchesNokeeVersionFromMainBuild() {
 		executer.withTasks("verify").build();
 	}
 }
