+++ subprojects/distributions-management/src/functionalTest/java/dev/nokee/nvm/NokeeVersionManagementServiceUsesVersionFromIncludedBuildParentFunctionalTest.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.init;
+package dev.nokee.nvm;
 
+import dev.gradleplugins.buildscript.blocks.SettingsBlock;
 import dev.gradleplugins.runnerkit.GradleRunner;
 import dev.nokee.internal.testing.junit.jupiter.ContextualGradleRunnerParameterResolver;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectory;
@@ -29,19 +30,30 @@ import java.nio.file.Path;
 import java.util.Arrays;
 
 import static dev.gradleplugins.buildscript.blocks.PluginsBlock.plugins;
-import static dev.nokee.init.fixtures.DotNokeeVersionTestUtils.writeVersionFileTo;
+import static dev.nokee.nvm.fixtures.DotNokeeVersionTestUtils.writeVersionFileTo;
 
 @ExtendWith({TestDirectoryExtension.class, ContextualGradleRunnerParameterResolver.class})
-class NokeeVersionManagementServiceUsesVersionFromVersionFileFunctionalTest {
+class NokeeVersionManagementServiceUsesVersionFromIncludedBuildParentFunctionalTest {
 	@TestDirectory Path testDirectory;
 	GradleRunner executer;
 
 	@BeforeEach
 	void setup(GradleRunner runner) throws IOException {
 		executer = runner;
-		plugins(it -> it.id("dev.nokee.nokee-version-management")).writeTo(testDirectory.resolve("settings.gradle"));
+		SettingsBlock.builder().plugins(it -> it.id("dev.nokee.nokee-version-management"))
+			.includeBuild("build-src")
+			.build()
+			.writeTo(testDirectory.resolve("settings.gradle"));
 		writeVersionFileTo(testDirectory, "0.4.2");
 		Files.write(testDirectory.resolve("build.gradle"), Arrays.asList(
+			"tasks.register('verify') {",
+			"  dependsOn gradle.includedBuild('build-src').task(':verify')",
+			"}"
+		));
+
+		Files.createDirectory(testDirectory.resolve("build-src"));
+		plugins(it -> it.id("dev.nokee.nokee-version-management")).writeTo(testDirectory.resolve("build-src/settings.gradle"));
+		Files.write(testDirectory.resolve("build-src/build.gradle"), Arrays.asList(
 			"def service = gradle.sharedServices.registrations.nokeeVersionManagement.service",
 			"tasks.register('verify') {",
 			"  usesService(service)",
@@ -53,7 +65,7 @@ class NokeeVersionManagementServiceUsesVersionFromVersionFileFunctionalTest {
 	}
 
 	@Test
-	void loadsNokeeVersionFromDotNokeeVersionFile() {
+	void fetchesNokeeVersionFromParentBuild() {
 		executer.withTasks("verify").build();
 	}
 }
