+++ subprojects/distributions-management/src/functionalTest/java/dev/nokee/nvm/PluginManagementDefaultRepositoriesFunctionalTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.init;
+package dev.nokee.nvm;
 
 import dev.gradleplugins.buildscript.blocks.RepositoriesBlock;
 import dev.gradleplugins.buildscript.blocks.SettingsBlock;
@@ -31,7 +31,7 @@ import java.nio.file.Path;
 import java.util.Arrays;
 
 import static dev.gradleplugins.buildscript.blocks.PluginsBlock.plugins;
-import static dev.nokee.init.fixtures.DotNokeeVersionTestUtils.writeVersionFileTo;
+import static dev.nokee.nvm.fixtures.DotNokeeVersionTestUtils.writeVersionFileTo;
 
 @ExtendWith({TestDirectoryExtension.class, ContextualGradleRunnerParameterResolver.class})
 class PluginManagementDefaultRepositoriesFunctionalTest {
