+++ subprojects/distributions-management/src/main/java/dev/nokee/nvm/ProviderUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.init;
+package dev.nokee.nvm;
 
 import org.gradle.api.provider.Provider;
 import org.gradle.util.GradleVersion;
